pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Joseph Maclean Arhin <josephmacleanarhin@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pip = "21.1.2"
kafka-python = "^2.0.2"
Flask-SQLAlchemy = "^2.5.1"
Flask-Migrate = "^2.7.0"
SQLAlchemy = "^1.4.11"
flask-marshmallow = "^0.14.0"
Flask-Script = "^2.0.6"
loguru = "^0.5.3"
flask-swagger-ui = "^3.36.0"
pinject = "^0.14.1"
apispec-webframeworks = "^0.5.2"
flask-mongoengine = "^1.0.0"
redis = "^3.5.3"
hiredis = "^2.0.0"
Faker = "^8.1.3"
fakeredis = "^1.5.0"
marshmallow-enum = "^1.5.1"
PyJWT = "^2.1.0"
cryptography = "^3.4.7"
Flask = "^2.0.1"
requests = "^2.26.0"
psycopg2-binary = "^2.9.1"
python-dotenv = "^0.19.0"
marshmallow-sqlalchemy = "^0.26.1"
blinker = "^1.4"
Flask-JWT-Extended = "^4.3.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.3"
black = "^21.5b2"
flake8 = "^3.9.1"
pre-commit = "^2.12.1"
Flask-Testing = "^0.8.1"
mongomock = "^3.22.1"
fakeredis = "^1.5.0"

[tool.black]
line-length = 89
target-version = ['py37']
include = '\.pyi?$'
exclude = '''
(/(
    \.git
  | \.hg
  | \.eggs
  | \.ini
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | .yml
  | .yaml
  # The following are specific to Black, you probably don't want those.
  | blib2to3
  | tests/data
)/
)
'''

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
